+++ README.md
@@ -3,10 +3,10 @@ Creación de una pagina web utilizando html, css y javascript( JQuery ) para el 
 
 
 
-### Recursos
-´´´
--API PRODUCTOS REPOSITORIO: https://github.com/juampi19/ProductosAPI
--E-COMMERCE DE JARDINERIA:  https://github.com/juampi19/ProyectoProgramacionWeb
--SERVIDOR DESDE HEROKU:     https://productos-api-rest.herokuapp.com/productos
--HOSTING DE LA PAGINA:      https://green-wood.netlify.app/
-´´´
+### Recursos:
+```
+API PRODUCTOS REPOSITORIO: https://github.com/juampi19/ProductosAPI
+E-COMMERCE DE JARDINERIA:  https://github.com/juampi19/ProyectoProgramacionWeb
+SERVIDOR DESDE HEROKU:     https://productos-api-rest.herokuapp.com/productos
+HOSTING DE LA PAGINA:      https://green-wood.netlify.app/
+```
