+++ formulario.html
@@ -22,7 +22,8 @@
             </div>
 
             <div class="col form-contenedor p-4 rounded-end">
-                <div class="text-end">
+                <div class="d-flex flex-column flex-md-row justify-content-md-between align-items-center">
+                    <a class="btn-all btn-all-mod mb-3 mb-md-0" href="index.html">Inicio</a>
                     <img class="logo" src="img/logo.png" alt="imagen logo">
                 </div>
 
@@ -65,7 +66,6 @@
                     </div>
 
                     <div class="my-3">
-                        <span>Regresa a nuestra Tienda <a href="index.html">Inicio</a></span> <br>
                         <span><a href="#">Recupera Password</a></span>
                     </div>
                 </form>
