+++ index.html
@@ -62,7 +62,7 @@
     <section class="hero">
         <div class="d-flex flex-column justify-content-center align-items-center h-100 gap-4">
             <h1 class="hero-heading">Green-World </h1>
-            <p class="fs-3">Lleva la naturaleza a tu hogar</p>
+            <p class="fs-3">Lleva la naturaleza a tu hogar, haz un aporte y recibe grandes beneficios</p>
             <a href="formulario.html" class="btn-agregar btn-agregar-negro">Inscríbete Aquí</a>
         </div>
     </section>
