+++ js/main.js
@@ -1,24 +1,12 @@
+import { validarFormulario } from "./validation/formulario.js";
+
 document.addEventListener( 'DOMContentLoaded',() => {
     initApp();
 } );
 
 
 const initApp = () => { 
-    var forms = document.querySelectorAll('.needs-validation')
-
-  // Loop over them and prevent submission
-  Array.prototype.slice.call(forms)
-    .forEach(function (form) {
-      form.addEventListener('submit', function (event) {
-        event.preventDefault()
-        if (!form.checkValidity()) {
-          event.preventDefault()
-          event.stopPropagation()
-        }
-
-        form.classList.add('was-validated')
-      }, false)
-    })
+    validarFormulario();
 }
